Remove old_name from kwargs when using IET helper.
The IET driver passes the input kwargs from export
directly to the IET driver, but one of the keys here
is specifically for migration and a bug associated with
going from nova-vol to cinder-uuid's.

This patch just checks in the IET code if we have the key set
and if so pops it out before passing through to iet.

Fixes bug: 1175207

Change-Id: I965bdfbe078d61b906aebc48961c1806a9fb0c59