Check that volume is at least minDisk size.
The Image service (such as Glance) provides a minDisk parameter to
indicate the minimum required disk size for the image. This change
ensures that the guidance is followed in the case of bootable
volumes.

Added unit test and fixed PEP8 issue.

Fixed issue with docstring.

Added a comment.

Fixes Bug #1175532

Change-Id: I505af53db70b5b286dbca2eaca4fc1f30a8c0b2a