Create an LVM utility to use for local storage.
This adds a simple LVM class for performing local
LVM operations.  The idea is that an LVM object
is instantiated based on a volume group, init
can be used to instantiate and query an existing
volume group, or create a new one if given a list
of PV's to use in creation.

See BP for information on where this is going and
how it will be used in Cinder and hopefully other
projects.

Implements blueprint: local-lvm-storage-utils

Change-Id: Iddde92af18f2317edc5f4583b2113c2b8117a4fe