Support mount options for NFS/GlusterFS volumes
This patch adds support for mount options for NFS/GlusterFS volumes.

To set options, append them to the line for the volume in the
nfs_shares_config or glusterfs_shares_config file:
 192.168.175.166:/testvol -o backupvolfile-server=192.168.175.177
 host1:/testvol2 -o backupvolfile-server=host2,rw,other_option

These options are used when mounting from the cinder-volume
service and the nova-compute service.  (There is an associated
nova-compute patch to enable the same.)

This is a useful change for GlusterFS because the
backupvolfile-server option enables a mount to succeed even if
the first server specified is offline.

Change-Id: I13bb0a930d6a05e803ccb50e6dc5d4b2b412ac11