solidfire: Add ability to override account prefix
The SolidFire account is created with a prefix based on the hostname
of the host running cinder-volume. This prevents cinder-volume from
being run in an active/active setup.

This patch introduces a new configuration option sf_account_prefix
to override the prefix used when a SolidFire account is created.

If sf_account_prefix is empty, no prefix will be added to
the SolidFire account name when created.

The default value is still the hostname.

Change-Id: I3b974789ea3e749cc8696b69754023ad76155179
Fixes: bug #1179671