Update/Publish volume service updates on delete.
The Filter scheduler weighing function was failing because
we were only updating the capacity info either on the periodic
60 second interval or after succesful creates.

The result was that if you exceeded your capacity within the
update interval you would no longer be able to create volumes even
if you deleted existing volumes until the update was performed and
published.

Since we're updating on on create, there's no reason why we shouldn't
do the same on delete for the applicable volume-service.

The periodic updates are still important for other purposes so they're left
alone, but with this change we keep things up to date so the filter scheduler
has the correct picture of the world.

Fixes bug: 1180976

Change-Id: Ic551f512c5a0cf51bafa8b60f0b6b7debbe22692