Remove execute permissions from test files.
Remove execute permissions from test_volumes.py and
test_storwize_svc.py, which caused the tests defined in them
to be skipped.

Fixing the permissions caused test_create_volume_with_volume_type to
execute, which fails. The cause of the failure is that
(1) LoggingVolumeDriver doesn't implement get_volume_stats(), and
(2) The fake RPC doesn't pass the statistics to the scheduler for
some reason.

Because this test doesn't test the scheduler, just change to use the
simple scheduler and work around the issue.

Change-Id: Ie01331d484c5c48c78b626eff1ad793dbbe7182c
Fixes: bug 1183416