Fix incorrect status for volume clone from image
If, when using a Ceph as a backend for Cinder and Glance, you
create a volume from a Glance image, the request is actioned
in one of two ways; (1) if Glance does not supply sufficient
information to Cinder, a full copy is made of the Glance
image or (2) Ceph uses it's copy-on-write capability to do a
clone. If a full copy is done, the volume is marked as
Bootable=True but if a clone is done it is not.

This does not appear to break nova boot from volume at present
but the status displayed by the cinder client is incorrect.

Change-Id: I2959fc3dfaf0e073781583019797bcd44f7789d6
Fixes: bug 1185533