Add missing attributes to xml deserializer for volume request
The v1 and v2 xml deserializers for volume requests are missing
support for the imageRef, snapshot_id and source_volid attributes
resulting in xml format create volume requests which specify any
of those attributes being silently ignored. This fix adds support
to v1 and v2 and extends the volumes unit tests to check for this.
Also includes some minor debug logging enhancements to make this
easier to find in future.

Fixes bug #1188581

Change-Id: Ib661c4a961c57e682e0e4e6db98d863b3a99cf71