Minor Logic bug in NFS Driver.
When checking oversubscription rate, we need to ensure that we
check the _allocated_ (apparant size) not the _available_ (actual
usage) on the NFS share.  This fix ensures we are now checking the
correct values.  This bug allowed (effectively) a one-time
volume allocation that would push the "apparant" size over the
oversubscription rate.

Change-Id: I189b689cb76ac2a243aa6e255ce0c80d86e247ce
Fixes: bug #1188877