Remove direct call to utils.execute.
The executes calls in the base driver class
are set up to use the exectue enginge passed in on
init.  There are a couple of methods that call
utils.execute directly, and as a result you end
up not using the expected execute engine when
you pass in something other than default for these
two calls.

We'd like to get away from the local execute altogether
and use the OpenStack common version only, so fix these
in driver.py and put the appropriate wrapper around them.

Change-Id: Ibc17c9c2a0c8f27b4533ed86377a66ce668aeab5