Content-length missing in put_object
Radosgw returns a "411 Length Required" in case the content-length is not
specified. The parameter content_type is not set which leads to the problem
that content_length cannot be determined. Since the content length is already
computed in backup/services/swift it shouldn't be a problem to set this
parameter.

Conflicts:
	cinder/backup/services/swift.py

Change-Id: I4c56823d70fcd1a801a75bfd14f474a2780dc1ac