SolidFire API RequestID is useless.
The SolidFire API can take an int or a string for a request-id,
when using an int however it needs to actually be a valid int.

The driver currently was doing an int(uuid) which is too large and
as a result get's truncated making it difficult or impossible to track
request-ids.

This patch chanes the generate request-id to form a hash of the uuid so
that it can be tracked.

Fixed bug: 1199198

Change-Id: I38d5e3d189e679678898c329a9fa173a47853d47