Refactor SSHPool.get() to use Pool.get()
In previous code of SSHPool.get(), we pasted the code
from Pool.get() and check if a connection is active before return
it.

However, it's much simpler and cleaner to just call the Pool.get()
and then check the connection before return. With this,we can free
ourselves from manually keeping up with code of Pool.get() in
upstream package eventlet

As a side effect,this patch fixed bug #1194393 which caused by a
previous bug in eventlet codes before revision 1072

fixed bug #1194393
Change-Id: Ic2bf2fa1ad82cf8669b6c491c955dcab39eb1510