Clone_image should return dict of vol properties, clone status.
The method will work fine in case of drivers not dependent
on volume properties like provider_location. It will fail to
restart in case of nfs drivers and also leave volume created as
result of clone_image functionality created in the
nfs share in dangling state after deletion as provider_location
is None. This fix requires dict of volume properties, cloned status
to be returned which facilitates passing back provider_location
in case of nfs drivers and hence resolves the issue.

bug 1200708

Change-Id: I590571e52d1c64b6dba7d7e76cd71badd74e51d1