Make unicode-to-utf8 conversion universal in ceph backup driver.
librbd does not like unicode strings. A patch was
recently applied (c98cf2d8) to make all config
values utf8 but there are other areas of the ceph
backup driver where this also applies. This patch
adds a new method which converts strings to utf8
and replaces all relevant uses of str() with this
new conversion.

Change-Id: I8c3b01d1ea00b601c599396e3a662f172ffb9757
Fixes: bug #1206044