Create key manager interface
This interface provides a thin wrapper around an underlying key
management implementation such as Barbican or a KMIP server. The key
manager interface is used by the volume encryption code to retrieve
keys for volumes.

This change is a copy-paste of the key manager interface accepted by
Nova. The major modifications are to the module imports (i.e.,
nova -> cinder).

Change-Id: I8f79165d66d67ad8faaca60946959c5a61811c59
Implements: blueprint encrypt-cinder-volumes