Update driver version to 1.1
This patch includes bumping the driver versions for both
Fibre Channel and iSCSI to 1.1.  Included a history change
that describes the changes since 1.0.

This patch also changes the way we do clone volume.  We no
longer have to check the status of the volume clone on the 3par
as it's instantly done now since 3.1.2 MU2 firmware.

Change-Id: I541ac7f97e6d1290bac6232d6c80e8ecf3a75dda