Move volume_clear and clear_size opts up to driver
The block_device driver pulled in the clear and clear_size
options for it's own use and broke the inheritance model needed
for multiple backend configs.

This change moves the two configs up higher in the heirarchy to
driver.py so that it can be shared/used by all drivers if they desire.

This means that they are now set properly in volume_opts.  At some
point we shoudl restructure this a bit to allow multi-backends to
specify independent settings for these configs.

Fixes bug: 1208964

Change-Id: I6afbdecea637f7e7d3417be537ab71fe97bffda3