Ignore purge_props for v2 Glance api and fix upload
The V2 Glance API image schema does not contain
purge_props. While this may be a bug in Glance,
we will ignore this property in cinder when
glance_api_version=2. This will not change behaviour
since Glance defaults this property to True, Cinder
never sets it to False and the v2 client appears to
ignore it anyway.

Also fixed image upload which is a seperate client
call to update in v2 (v1 update does both).

Change-Id: I0ba1d7d920984cface57795ace160ec300ff75e2
Fixes: bug #1210467