Fix handling ImageUnacceptable in create_volume
In the create_volume flow ImageUnacceptable exceptions are
not properly handled.  _copy_image_to_volume can receive an
ImageUnacceptable exception from copy_image_to_volume if fetch_to_raw
is used.  Currently the ImageUnacceptable exception is changed to
a generic ImageCopyFailure exception which does not make the cause
of the excpetion clear.

This change adds handling and raising ImageUnacceptable exceptions
in _copy_image_to_volume.  It also adds the exception to
no_reschedule_types as it doesn't make sense to keep retrying the
copy if the image was found to be unacceptable.

fixes bug 1209021

Change-Id: I9aacf23dfaaa1a4df0a869d1c9544c376488f01b